mensaje de confirmacion antes de eliminar

+++ resources/views/usuario/edit.blade.php
@@ -9,7 +9,7 @@
 {!!Form::submit('Actualizar',['class'=>'btn btn-primary'])!!}
 {!!Form::close()!!}
 
-{!!Form::open(['route'=>['usuario.destroy',$user->id],'method'=>'DELETE'])!!}
-  {!!Form::submit('Eliminar',['class'=>'btn btn-danger'])!!}
+{!!Form::open(['route'=>['usuario.destroy',$user->id],'method'=>'DELETE','onsubmit' => 'return confirm("seguro que quieres eliminar el registro?")'])!!}
+{!!Form::submit('Eliminar',['class'=>'btn btn-danger'])!!}
 {!!Form::close()!!}
 @endsection
